Allow use from ClojureScript as well as Clojure
This changes the source files to cross-platform with reader
conditionals. Almost all functionality works in ClojureScript as well
as Clojure, the sole exception being XML parsing, which relies on Java
libraries for which I knew of no drop-in replacements.

This also establishes a non-snapshot version so the project can be a
dependency of other projects which are being published as actual
releases.